Enable client file locking
The Persist package now provides a new interface, `Persist.Lock`.

`Persist.Lock` provides a simple `acquire` method for obtaining a lock.
Each lock has an ID.  It's up to the API user to decide how to name
locks, and to make sure that they are released at the right time.

As currently implemented, if the application crashes, all locks will
automatically expire after 2 minutes, but API users should still make a
best effort attempt to release locks explicitly when (for example) a
window is closed.