Improve robustness of HCR snapshot restoration
Previously, state injection would stop at the first error.  This is no
longer the case.  A list of errors is accumulated throughout the
injection process.  Any error will prevent state injection from
proceeding down that branch of the render tree, but will not affect
other parts of the page.

E.g. if a toolbar component has been changed, causing an error, and some
of the toolbar's children have their own state, the toolbar's children
will not receive any state injection (because of the error further up
the tree).  However, components later in the page (below the toolbar)
will, in most cases, still receive their state injection.

Aside: in most cases, the components with state are actually pretty far
up the tree (e.g. the PlanTab component, or the ClientFilePage
component).  This is fortunate, since it means that they are among the
first to receive state injection, and therefore are unaffected by errors
farther down the tree.

/cc @applebya