Abstracted listeners to main.coffee, page & timeout listeners
> main.coffee on init calls the pageComponent’s public
registerListeners() method, which returns an object of event-name and
action key-value pairs, which are iterated over to create each Backbone
event binding.
> Refactored the timeout component to return the same kind of listeners
object as well, which is tacked on to @pageListeners
> If pageComponent.registerListeners() returns `false`, that means we
don’t want ANY listeners (including timeout) registered - for example,
on the loginPage.
> registerListeners() must be set up on every pageComponent (it’s
asserted in main.coffee), even if it only returns an empty object or
false.
> Listeners are unregistered after the React component is unmounted,
using the same @pageListeners obj built in registerPageListeners()
> Removed all mention of @state.isClosed (and isClosed in
timeoutDialog) since this stuff can’t exist anymore after the
pageComponent is unmounted.
> Reviewers @timmclean @conceptualspace