Impl Atomic module
Normally, file operations involve many steps, and could fail part-way
through.  The functions in this module are designed to act as if the
entire operation is only one step -- i.e. it either succeeds 100% or
leaves no trace that the operation was attempted.  This helps prevent
application crashes and network problems from corrupting user data.

Needed for issue #1.