Update persist object indexes after creating new revision
It was improper to update the indexes before the new revision actually
existed, since the revision files are the single source of truth.  Now,
in the situation where the second part of the transaction is blocked,
the indexes will be out of date, but will not contain data that is not
part of the object's revision history.

See issue #1.