Better tooltip configuration to return unscaled values
> Rather than hijacking the tooltip.content every single pageLoad
(which had a LOT of weird bugs come out as a result), I’m explicitly
re-defining the function in within @_chart.generate() . Works on
first-run, and much smoother.
> Had to explicitly use Chart.dataSeries rather than @dataSeries.
Mainly because the tooltip.content function needs “this”/@ as well.
Might be able to refactor this a little better, but it works.
> Will eventually need to clean up some more of this code (copied from
https://github.com/masayuki0812/c3/blob/master/c3.js#L3807), but that
can wait until we have to do any serious work on the tooltip structure.

> Ref issue #42

> Reviewers @timmclean @conceptualspace