Refactored analysisTab -> generate once, show/hide metrics
> So there’s not a whole lot you can actually do with a C3 graph after
you generate it, and try adding new stuff into the .load() params. To
mess around with things like tooltip, x-ticks, etc - you need to use
generate(), because hacking it further downstream was REALLY painful
and unreliable.

> I took a bunch of the business logic out of the view component, and
into Chart. This includes the actual generation etc, because that needs
to run after all the data processing completes (as opposed to before)

> All metrics are hidden by default when the chart’s generated. I
modified the `_updateMetricSelection()` handler to show/hide the metric
in the chart. It needs a little work, but it’s functional!
>> To prevent the legend and metric checkboxes from falling out of
synch, I’ve disabled any clicking the legend.

> I’m thinking of pushing the x-ticks issue to the next milestone
(perhaps 1.2.1?) in favour of communilatively grouping metrics by day,
and setting `tick: {fit: true}` to prevent the onslaught of x-ticks.
Can still make it only show chart if >= 3d of it. David agrees that for
now it’ll be good to group metrics by day.

> Reviewers @timmclean @conceptualspace
> Ref issue #41