Lock is released on timeOut, re-acquired on pass confirmation
> Solved my problem @timmclean with the event listeners. When I merged
in timeoutListeners on main.coffee, it was overwriting any of the same
listeners I had in pageComponent. For now I’ve just separated these,
which works fine. Ideally I’d like to refactor this into sets of dupled
arrays, so the same key can exist more than once. Or maybe just divide
by set type, ex: `{timeout: {}, pageComponent: {}}`

> Had to create a fake persist/eventBus object for the lock test. I
don’t understand how to generate an actual test session, because I
would need this in order to test the forceStop functionality in
@_acquireWhenFree

> Fixed up clientFilePage/index.coffee ‘timedOut’ listener to only
erase the clientFileLock once released, and only even bother with all
this if the window even has a clientFileLock.

> Reviewers @timmclean