LockOp state + release(), @_killLocks() method
> Refactored all lock releasing into a single @_killLocks() method,
since it’s needed from both deinit() and the timedOut event.

> @acquireWhenFree method now synchronously returns an object
containing a cancel() function. This sets `isCancelled` to true,
effectively allowing the Async.until() to return. This way we don’t
have to rely on listening to outside events, which is much more elegant
(thanks Tim!)

> @acquireWhenFree interval now optional, set to 1s by default

> Cut down the interval time for npm test, so it doesn’t waste time
during testing.

> Removed console logging

> Much improved UI for the Read-Only Notice. Also recognizes when same
user has the the same file open.

> Removed old ‘lockAcquired’ event, this is leftover code from older
implementation of @acquireWhenFree

> renderName() method in utils was pushing in an empty middle name (so
an extra space when joined), fixed!

> Reviewers @timmclean @conceptualspace