Events appear in progNote list and each note in detailView
Remodelled progEvents to be a child of clientFile, so now much easier
to access this event data through the clientFile itself.

@timmclean @conceptualspace @dgotlib
This means that there now needs to be an ‘progEvents’ directory within
each clientFile directory. Here’s a bash script I wrote that you can
run from the root directory to recursively create these folders for
you… otherwise I think the app might crash for you when saving a new
progEvent: `for d in ./data/clientFiles/*; do mkdir $d/events; done`

OR you can just delete the `./data` folder and start over. Your choice
:)

> Corresponding events appear at the bottom of a progNote (left pane)
and an entry from a progNote (right pane). This works on both the
clientFile page, and the newProgNote page.

> Reviewers @conceptualspace @timmclean