AnalysisTab displays events on chart
> This works fine, although still need to write an algorithm that sorts
the events into rows, so like:
_________ _____
 ____________________

> @timmclean I’m SO close to making it possible to giving these `rect`
tags some key-bindings that make them show the eventData. I’ve attached
the progEvent ID to the `rect` as a class, which works fine. My problem
is getting the `.c3-regions` group to display on top of the `c3-chart`
one, because otherwise I can’t even click/hover on them.

1. I first made this work by doing
`$(‘.c3-regions’).insertAfter(‘c3-chart’)` Problem is that this seems
to steal it out of the `@_chart` context, so when it re-renders in
anyway (such as `@_chart.hide()`), everything falls apart.

2. What I have in place now is cloning `c3-regions` instead, so the
original context is still there. But this displays nothing…

Any help on this would be really appreciated, it’s a pretty essential
part of the analysisTab .

> Reviewers @conceptualspace @timmclean