Refactored timeSpan to include progEvents
> Compiles an ordered list of all unique days (beginning of day in unix
ms) containing a metric timestamp or progEvent start/end timestamp.
This is now used to calculate the dateSpan, which is used by the
dateSpan slider and @_chart (on initial render)

> Next challenge is to get this timeSpan to update the UI accordingly
when any new metric/event added falls outside of the given timeSpan.

> Oh, and added some more region rows just in case. Still need more
testing on these, as rows with an index of 3 or more tend not to show
for some reason…