Write v1.3.0 -> v1.3.1 migration code
FYI: @conceptualspace
Reviewer: @applebya

I created a new migration module to hold code that upgrades the data
directory from one version to another.  The idea is that, when deploying
an upgrade, we would open the dev tools, and run something like:

    M = require('./migration')
    M.runMigration('data', '1.2.0', '1.3.1', 'user', 'pw')

This will figure out what migration steps are needed (e.g. v1.2.0 ->
v1.3.0 -> v1.3.1), and modify the data directory as needed.

This upgrade was particularly complicated.  I redesigned how user
accounts are stored/encrypted on the network drive so that admins can
access any account (to reset a password, etc).  Since that's not
currently the case, that made upgrading difficult... since as an admin,
we can't currently see all of a user's information.

I added a bunch of temporary code wrapped with "BEGIN v1.3.1 migration"
and "END v1.3.1 migration" once the migration is complete, this code can
be removed (hopefully by v1.5.0).  This code upgrades user accounts, one
at a time, as they log in.

This won't be needed next time, now that key escrow is in place (and I
learned my lesson).

More details about this will be over on #198, key escrow.