clientFile data set into state all-in-1, removed _handleDataSync (temp)
> Fixed renewAllData to assign everything to state all at once at the
end, instead of 1-by-1. @conceptualspace, remember the weird bug we
were getting from planTab on Friday? This should resolve it, thanks to
blessings from the Oracle.

> Removed `_handleDataSync()` for now. See issue #273, we’ll continue
discussion in there on how to handle this stuff. We’ll focus on testing
this tomorrow to make sure it won’t explode in the meantime.

> Ref issue #273
> Reviewers @conceptualspace @timmclean