clientProgramsDialog can display and create programs
Initial UI and data logic for programManagerDialog
> Displays all programs in table list
> Can open new dialog to create a program

TODO:
> Restore loading render flow
> Modify (revise) program… I’m not convinced that we should even have
this collection mutable for performance reasons.
> Delete program (choose which program to move clientFiles to)… I see
this as being tricky, since if a clientFile’s locked, we won’t be able
to modify their data. Perhaps a solution might be to simply store
clientFileIds in each program instead, and have that be the only source
of truth… thoughts @timmclean?
> Assign clientFile(s) to program(s)… will encounter same locking
problem as above.

> Reviewers @timmclean @conceptualspace