Refactored admin control into component, clientFileManager working
> clientSelectionPage now includes managerLayer within its `render()`,
so we can pass all kinds of useful data into it

> Menu options pull up this managerLayer, with the specified manager
component. Currently passing all needed data across all components for
now.

> Continuing to use Dialog for creating a new client. We might want to
eventually work this into the UI without requiring a dialog, but I like
how it focuses the user on one thing - so probably better this way.

> Created a DataTables-like component for an OrderableTable. Pretty
flexible for now, but needs more work. The plan is to integrate User
Accounts and Programs in with this later today, saving me a lot of time.

> Programs option breaks the app, and User Accounts I still need to do
too.

> Ref issue #278, #305
> FYI @conceptualspace (still a little more work to do, but have a go!)