data dir now internal to .app on mac builds
> on mac, having the data dir and .app inside a “konote” folder made it
difficult for users to find and launch the app. this was done
originally to make it easier to find the data dir when upgrading and to
support sharing the data dir between multiple users

> now moved the data dir inside konote.app - should simplify
installing/launching/finding the program.

> when upgrading, users should be instructed to first copy the data
folder from the cotents of the .app and then paste into the new version.

> sharing should still be possible by creating a path to the contents
of the .app file

> TODO: update readme

fyi @applebya @dgotlib