Create program, better table layout, colorPicker, orderableTable improvements
Note: `npm install`

Programs
> Very improved UI, includes color keys, and only shows what’s needed
> Removed TinyColor completely
> Chose new colorPicker (spectrum) which includes both a flexible
picker and a bunch of preset colours on the palette. We can refine this
list later on.
> Can create a new program. Simplified UI by having colorKey appear
beside the name

ProgramLinks
> Changed schema to include a `status: (string)` property, rather than
a boolean isActive. We’ll have a list of predefined statuses indicated
here in the schema. Will document these when complete.
> Began on refactoring link management dialog

OrderableTable
> Lots of refactoring to enable more functionality, and also radically
simplified the API.
> OpenDialogButton component simply extends in the props, lets
OrderableTable handle more of the business logic.

Schema
> Decided finally on using & storing the Hex# format for color key.