programLink status schema, pre-defined values
> @timmclean @conceptualspace @dgotlib, this is how Im envisioning the
schema for clientFile-program links. Eventually we’ll add a new status
for say ‘graduated’ and ‘expelled’ as use-cases come up.

The current UI for adding/removing clients to a program only deals with
the actions of enrolling/unenrolling clients, so I’m keeping it simple
like this. Eventually we’ll have a way for the user to indicate “he
graduated”, “he got expelled, here’s why”, etc…