Store and validate info about where each object belongs
This solves a security problem that allowed an attacker to copy a
persistent object from one collection to another, without the
application detecting the tampering.  The solution was to store
information about where the object belongs inside the object file
itself, and validate that information whenever the object file is read.

This change requires a migration that updates every object in the
database.  To fix your existing data directory, run the 1.4.0->1.5.0
migration in the usual way
(`require('./migrations`).runMigration(...)`).

Closes #191

Reviewer: @applebya
Reviewer: @conceptualspace