Completely refactored state & submission logic for programManager
> So I this is my 3rd and final refactor on the logic, I think I’ve
finally gotten it right :)

> Note: If you’re running this for the first time, create a
“clientFileProgramLinks” folder in /data.

> [un]enrollment is handled by private methods. I realized that
@state.clientFileProgramLinks needs to hold ALL links, irregardless of
whether they’re enrolled or not. This really simplifies the save process

> I have this `_originalLinks()` function which pulls a stripped
version of the @props.clientFileProgramLinks, which holds the same
number of properties as the current @state.clientFileProgramLinks for
easy diff-ing. This gets called once more on submit, to ensure we’re
working with the latest version of it that the app knows about.

> Ref issue #278
> Reviewers @timmclean @conceptualspace