Migrates progNote format (#7)
> Finished migration for #7. Took some time to refactor it into a
version-specific utility, so the actual `run()` is easier to understand.

> Added some more clear console.logging, and info steps . The only
other console.info is for the new/migrated progUnits, which I need to
inspect closely during testing.

> Had to account for notes that occurred before a section was created.
For example, I create a progNote that has 2 sections, and then create a
new section on my plan. Since that progNote only has 2/3, the spare one
shouldn’t be mapped over.

> I’d like to have “read progNote” (line 351) as part of the Async
series below, but I’m a little confused about how to elegantly skip the
rest of the series (line 360) when the progNote doesn’t qualify for
migration. Perhaps instead I should extract the series out as a
function?

> Ref issue #7
> Reviewer @timmclean
> FYI @conceptualspace