Force file extension during writeFile
> @timmclean, this is our jimmy-rig solution for #370 where it’s
inconsistent whether or not the file extension is included when saving
the file. We see this mostly on Mac, but there’s some weird things in
Windows as well. Anyway, we decided to force the file extension on to
the path at the very end, to ensure it’s there. This is a known nwjs
bug.

> Also moved final metrics save step into end of series