Extracted chart, analysisTab can filter by plan target/section
> All targets are selected by default, so we only need to hold the
excludedTargetIds (unselected targets) in @state.

> Line70, needed something that could nicely tie targetIds in with
their metricIds, so it turns out looking like `{targetId,
[metricIds]}`. This is used later in @_metricIsExcluded which
determines whether or not the metric should stay in selectedMetricIds
during the render.

> A disabled/excluded metric from the plan filters is shown by a
strike-through on the metric name. This is to indicate that this metric
may not be toggled, as it’s currently filtered out from the plan
filters.

Issue #365