readLatestRevisions for progNotes, progEvents, metricDefs
FYI @conceptualspace

> Mutable dataModels require `readLatestRevisions()`
> Avoid using `event`, as it’s a reserved variable name. I’ve tried to
only use `progEvent` around the app for consistency and to not confuse
it for say a backbone or click event. That said, it might’ve been me
who did this, lol!
> Added filter for ‘cancelled’, since by default we don’t want to
provide progNotes that were created mistakenly. Opened #404 for
continued discussion.
> I’m not sure about exporting full histories of progNotes, we’ll wait
for the need to arise.

> Still experiencing the double confirmation window after exporting
metrics/progEvents. Reopened #385

Issue #403