New Atomic method: writeJSONToFile, deactivate() and setPassword() now use it
#316

FYI @timmclean @conceptualspace @jophis

I needed a shorthand atomic method for writing to a file as plain JSON,
as opposed to a `new Buffer()`. This is used in `Account.deactivate()`
and `DecryptedAccount.setPassword()`. If this operation were to fail
prematurely, then the user account in question would be f*ed.

Also added tests. Still might need a test for the isValidJSON util,
which I got from http://stackoverflow.com/a/20392392/1694975