Refactor loading seq for: login->clientSelection, ->clientFile, -> newProgNote
Using our eventBus to listen for and trigger custom events, to only
show the window once everything’s completely ready. The loading spinner
on the original window persists until it hears this event.

I’m still experiencing a bit of a flicker, so I’ve delayed the final
loading trigger an extra 500ms at `componentDidMount()` in the UI of
the new window.

#464