Remove ReactBootstrap factory build step from main, use on-demand from module
Until React-Bootstrap add in support for factories (aka non-JSX users),
this module generates a factory for pre-defined components, which are
passed in as args.

This removes the build step from main.coffee, which ran for the new
window regardless of whether ReactBootstrap was being used or not.

This feature was added (briefly) to ReactBootstrap here
https://github.com/react-bootstrap/react-bootstrap/pull/536 but got
pulled out of the release. When that DOES happen, it’ll look something
like this:

`{Dropdown, MenuItem} = win.ReactBootstrap`
and we’ll just have to load a different path in its <script> tag in
main.html

#475