Unregister window close listener before navigating
I'm 90% sure that this is needed.  `navigateTo` isn't currently used in
our codebase, so I think that's why this hasn't blown up on us yet.

When the first page loads, it sets up a close event listener on nwWin.
When navigateTo is called, the page URL changes, but the same nwWin
stays.  The new page registers another close event listener on that
nwWin, but the old listener is still there.  If a close event occurs,
then the old listener will still be called.  This could cause problems,
because the DOM and all of the React components of the old page are no
longer valid... weird behaviour may follow from here.

Reviewer @applebya