Persist checks for invalid username, calls back with CustomError if cb provided
Originally we had checks on the front-end for username validity, but
any failed check on the back-end would throw a catastrophic error.

Now, it uses the parent function’s callback to return a `new
InvalidUserNameError()`, which is then handled on the front-end. This
reduces a bunch of code on the front-end, and we no longer need to
export the userNameRegex.

#457