WriteFile test for local app directory, at newInstallationPage
Tests to make sure we can write a file to the local app directory. This
accounts for when someone tries to run the app before unpacking to
their Applications folder (Mac users)

We’ll need to sit down and figure out all the possible cases, and
handle them informatively so we can help the user debug.

Similar error handling will need to be done with actually creating the
data folder, and making sure that it’s available.

FYI @conceptualspace @jophis

#454