Target history shows context of current progNote
So this is something I just felt like trying out while on the bus. I
noticed that while browsing through progNotes, it was a mental
challenge to match the history on the right pane with its parent
progNote on the left pane.

Rather than drawing the user’s attention too much, I went for 2 shades
of a bright yellow for the background of a history item. It goes to the
darker shade when the user hovers over the actual target note entry.

Best you just try it out, and see what I mean :) It’s on a separate
branch, will merge into develop if there’s no protest.

FYI @conceptualspace @jophis