clientFileProgramLinks now being created
@applebya @conceptualspace woo!
moved the create link function into the callback of the create
clientFile function (since one depends on the other anyway) and it
solved my issues. I couldn’t set state within a callback so i moved the
whole next function into the CB for create clientFile. Things are
finally working! If you want to test it, click the plus icon from
landing page and only click a program once. clicking it twice adds its
ID to the array twice and that is a nono. Next step is to make the
buttons toggle to avoid this problem and clean up the front end. Thanks
for your help.