Remove duplicate async call of @ _generateCurrentTargetRevisionsById
#541

Basically, I had set up another @props listener to check for changes in
the planTargets, because we needed to know when an individual
planTarget was updated (status changed) and update our local working
state to match.

Thing is, these both ran pretty much the same asynchronous @setState
call when both changed at the same time - i.e, when creating a new
target & section. That means they overlapped, and likely corrupted the
state somehow, later leading to Immutable methods (‘update’ in this
case) choking on it.