Resize textareas on window resize
In #576, the textareas were too tall for the content. This was happening
because the textarea height was being determined before the window is
maximized. The height was correct for the smaller window dimensions, but
incorrect once the window was made larger.

The fix: automatically redetermine the height whenever the window size
changes.

Closes #576