Prevent ExpandingTextArea height glitch on resize
During the textarea sizing calculation, the textarea height is set to
zero. If the targets list div was scrolled to the bottom, shrinking the
textarea would cause the scroll position to move up. When the resizing
completed, the scroll position would no longer be at the bottom. This
created some weird scrolling in some cases when the user types in the
textarea.

This fix wraps every ExpandingTextArea with a div. This div is used to
prevent the ExpandingTextArea component from shrinking to zero height
during a sizing calculation. This makes the scrolling behaviour more
sensible.

Probably fixes #529

Reviewers @applebya @jophis