Add react-devtools with ImmutableJS support (fork)
@conceptualspace @jophis

Now we can use React DevTools to debug our components, and just as
importantly, all ImmutableJS structures are automatically converted
(`toJS()`) for you. This provides a precise view of all props & state,
and lets you play around with it.

The extension is a fork of https://github.com/roblabla/react-devtools
(with Imm functionality), and a small customization to its
`manifest.json` which adds the permission for the url prefix
`chrome-extension://*`.

I tried doing the same thing for the React-Performance chrome
extension, but it puked - so we’ll have to benchmark performance the
good ol’ way. More on that coming soon…

To use: open the current window’s devTools with Ctrl+Shift+J, and
you’ll see the React tab get added. Hit that, and start looking through
the React components’ props & state!

I think this will work fine in /scripts, but we can consider making
this a separate repo. Let me know your thoughts.