Use startTimestamp for ordering
@dgotlib, another detail I need your input on:

In the progNotes tab, we list all progNotes by date they were created
(or backdate if exists).

Now that we mix in globalEvents, I believe the date used should be the
startTimestamp of the globalEvent - rather than the timestamp of when
it was created in the DB.

It would make more sense contextually for the patient care history.

#124