Change clientFileStatusIndex decrypt arg to 6
#552 @jophis

In the case of indexes, we’re dealing with the name of the containing
folder - not each revision.

We only need to loop over the folder contents/revisions when adding a
new field to the dataModel, like we do in `addClientFileStatusField`.

Here’s how to think about the layers of data in the file system:

```
/ClientFiles
- /Data Folder: [index1, index2… , ID]
- - /Revision File: [ID, revisionId]
```

The reason why it was kind of working for you before was that the var
`clientFile` coming from the first forEachFile call was being used in
the second one. So it was modifying the folder indexes properly, but
then saving it as something like `clientFiles/clientFile/clientFile`.

With all this in mind, take some time to go over the migration before
these changes, and after, to see how this maps to the folder structure.

As we thought, you were 99% there :)