Changed back to diffing by characters
@dgotlib, since we came to the conclusion in one of our meetings that
most modifications to a progNote will be to make minor corrections,
then we will diff the individual characters instead.

Eventually we’ll add in a diffing button for planTab again #630, which
I assume will require diffing by words. We can look into making this
diffing smarter, so it can figure out which kind of diffing to use.