Refactor inactive client filter to use results
The filter for inactive clientFiles needs to reflect the inactive files
within the current results, whereas it was using
@props.clientFileHeaders .

Now it updates in real-time with the results. This kinda causes the
layout to jump up and down a bit, but we’ll fix this with a dedicated
Filters area soon.

FYI @jophis