Reverted 321e9bbd01322c7fb72fcc25f7a89f4281401e50
Turns out that blurring away from the metricLookupField forces the user
to click “Add Metric” over and over again.

In favour of simpler input for adding metrics, we’re going to leave
this as-is. Will create a new issue if this becomes one, otherwise it’s
just something I noticed as looking odd - but turns out to be better.

#683