intelligent diffing
> i'm sneaking this into 1.8 but happy to demote to develop branch if
the team votes to. it was just too cool to resist.
> basically we can make the diff smarter by running it in 2 passes. in
the first pass we just want to figure out the number of substitutions,
or how similar the original text is to the new text. then based on that
we just do the diff again -- but if the similarity was high we diff by
word, or if the similarity was low we diff by sentence
> i used the DiffMatchPatch library for the first pass, only because it
allows us to specify a timeout (in case it gets too costly we can lower
the timeout and guess which second method to use based on what was
processed in the allotted time). otherwise the same approach would work
with jsDiff
> should resolve #618
> fyi @applebya @dgotlib @jophis -- please test