DialogLayer component to wrap table, handles layering for dialogs
Eventually this might become a top-level component, so we load dialogs
in a more consistent way. Anyway, this wraps around a table and
provides an `open` method for dialog.

Any dynamic props need to be loaded into the main statement of
DialogLayer, so only static IDs should be loaded into the `open()`
props.

#655