Removing 'overflow: hidden' allows for multi-page printing
@conceptualspace this is what was causing #705.

Since it didn’t prompt a scrollbar in
printPreviewPage, it wasn’t able to print the contents that flowed over
the window container.

Problem here is that now when you go to load a new clientFile, there’s
a scrollbar beside the preloader. One of them must have some padding,
that’s causing the overflow on a blank window.

Need to find a solution that satisfies both…