Override standard body margin, prevent scrollbar
I remember having to do this in the days before CSS resets & Bootstrap
lol!

`body` has a bit of a margin by default.

Since it takes a little longer to show the styles for us (especially in
dev), this margin still holds true. So have to manually override it.

Also trying to make up my yearly quota of semicolons… :)

Fixes #705 @conceptualspace