Temporary fix for <th> padding when hidden column present
Opened issue
https://github.com/AllenFang/react-bootstrap-table/issues/628 with
react-bootstrap-table .

This is a temporary fix for this table’s headers getting slightly
displaced when a hidden column (`hidden: true`) is in the table body.

Will probably submit my own PR for this, but apparently the owner’s
going to have a look. This fix will do for the meantime.

#655