Add dataIsReady state for managerLayers that load data, fades in table
For managerLayers that load data, I’ve added a bit of a fade transition
by not rendering the table until everything’s loaded in.

We might be able to make this look a little better by showing the table
<th> headers straight away, and then fading in the rows. Till then, at
least this solution prevents all the jarring transitions between
managerLayer tabs.

#655