Fix basicUnit display of authorProgram
Found while working on #640

The colorKeyBubble was not displaying the authorProgram in the
progNoteDetailView.

For example, if you look at the history of a target, it shows all the
colorKeys. If you look at the history of “Notes” (hard-coded
basicUnit), then it doesn’t show.

Updated basicUnit to extract the userProgram as well.