Replace 'hiddenColumnFix' with 'rightPadding' fix
What I wanted to do here was apply padding-left to the first visible
column, and padding-right to the last visible column.

Problem is that last-child gets kinda wonky when it comes to
differentiating by className. For example, we’d might dismiss the
hidden one in CSS using `td:not(.hiddenColumn):last-child`. That works
fine for first-child, but not last-child unfortunately.

Anyway, this (temporary) solution is more robust, and provides a more
consistent layout. Solving
https://github.com/AllenFang/react-bootstrap-table/issues/628 will let
us remove this fix.