Fixed faulty openNewProgNote logic
So the `@setState {isLoading: true}` triggered at the beginning wasn’t
getting set to false again, even if the user cancelled out of the
openNewProgNote process.

If the plan had changes, and I opted to go save the plan, the “New
progress note” button remains disabled.

Yeh, this kind of leaves an unresolved Async.series, but at least it
performs as expected. Suggest we refactor this at some point.