ClientAlerts component
#736

Since clientFilePage code is massive, extracted this out to a component
to keep things a little more modular & reasonable.

We assume for now there will only be 1 alert, as a simple textarea
(content), which is initially created, and updated.

As a full feature, this will be an itemized list of individual alerts,
so the dataModel is modelled on that eventuality.

Automatically creates a quickNote.