Fix change account type bug
Promoting a user account to admin wouldn't properly insert the system
private key into that user's account data, leading to a crash if that
newly admin'd user tries to reset someone's password, for example.

The problem here is that the privateInfo object must only contain things
that can be JSON.stringify'd. Therefore, the systemPrivateKey field
cannot contain a PrivateKey object, it must instead contain an exported
version of the key.

Reviewer @conceptualspace