DialogLayer for userAccounts should use raw @state data
#756

FYI @conceptualspace

Fairly simple fix. At the top of the render, we do some data
manipulation to suit the needs of the table. But when we open up the
Modify Account dialog, we want the raw data - not the bastardized
version.

I recall fixing this on a few other ManagerLayers, but this one must’ve
been missed. This is likely the same bugs in the metrics one, will tag
you.