Use tableData for metricManager table, @state for dialog
Same fix as for #756

Completes #755

FYI @conceptualspace

Added back in original code, which was needed to display “active”
instead of “default” on the table.

To prevent confusion, and maintain consistency, I renamed
`metricDefinitions` to `tableData`. This separates concerns a bit
better.

Also changed the button text in the dialog for “Default” to “Active”.