Shrink font-size by 15% when number of plan items > 16
#80

First attempt at making the condensed view fit a little better for plan
that has lots of items.

For now, I’m assuming the breaking point is 16. This is total
(including inactive items, even if not currently being shown)

We can improve on this later, but the basic functionality is there.

FYI @dgotlib