Unignore develop.json (enable shiftSummaries), disable it for default (generic) build
FYI @jophis @conceptualspace, since we’re not really using the ignored
develop file, we might as well un-ignore it and add a view things for
development. This gives us the overrides we need for development (such
as having shiftSummaries feature), and the default.json holds all the
config for the generic build.

If you need to do any wacky stuff with develop.json that’s just for
your own purposes, simply don’t commit it :) That’s what I’ve been
doing with default.json anyway .