Replace remaining components' err handling with handleCustomError
Need to start forcing errors around the app in order to thoroughly test
this.

In that case, since we’re so close to release, I’m going to keep this
on a branch and keep it up-to-date w/ develop. Need to consult with Tim
on the implementation details as well.

#713