Adjust timestamp format for space (now: 26th Sep, 2016 - 6:43pm)
Was: September 26th, 2016 at 6:43pm
Now: 26th Sep, 2016 - 6:43pm

Interface now holds up down to 1100px wide. I think we need to get this
down to 800px, and simply set that as the max window width - at least
for the ClientFile window.

Any alternate ideas for timestamp format are welcome. BTW, this format
is entirely configurable on a by-customer basis.