openWindow arg 'options' defaults to empty obj
`if options.maximize` was puking, when options wasn’t defined.

We may want to make ‘options’ an entirely optional argument. So rather
than having to do `openWindow(PARAMS, null, CB)`, we can just do
`open(PARAMS, CB)`

Tim taught me out to do this. Just check to see if options is
instanceof a function, and if so, `cb = options`.

@conceptualspace