InfoTab full style/layout refactor
Managed to figure out a layout that works and preserves space a bit
more, with the buttons only appearing as needed (without messing up the
layout).

Since just “Information” as the tab name may not be enough (I see where
@jophis was going with this), I figured a page title indicating “Client
Information” might help be a bit more explicit.

To get the aligned form groups, I used a simple `<table>` approach,
which keeps columns in sync.

The button toolbar is completely hidden, until there are changes - then
it flies in from the right side when changes are detected. I feel this
keeps the layout sparse, if the user’s only in there to read.

Added in an avatar, which would eventually be the client’s photo. And
hey, since we’ve figured out saving files to the DB, maybe this feature
isn’t so far off…

Oh yeh, and I switched out textarea for our now infamous
ExpandingTextArea component.

That’s about it. I know it looks like a lot of changes, but I kept a
lot of Jo’s original code, and just happened to get on a roll…

#114