Star/highlight functionality for Event Types in analysisTab
#817

Uses an inline `<style>` to fade out any progEvent regions that don’t
have the typeId in its className. This is handled by
`InlineHighlightStyles()`

We have a state for transiently-highlighted eventTypes (onMouseEnter),
and one for persistent ones - which is what the star signifies.

Now that it’s on there, I actually really like the star. :+1:
@conceptualspace