Don't factor in empty endTimestamp to xTicks calculation
At this point, I’m fairly sure most of the crashes we’ve experienced
lately in analytics revolving around timeSpan have been caused by this
simple logic error.

Basically, when we first decided to allow for point-events to be
created,  analysisTab never got the memo that endTimestamp can be empty
(it’s `.allow(‘’)` in the dataModel). Downstream, this would cause the
xTicks to contain a `NaN` value, which caused havoc further down in the
render.

@conceptualspace I’ve tested through this, creating new data etc - no
problems, no more rabbit-holes. Suggest we c-pick to release.

FWIW, this could be the root cause of #814, and the call stack overflow
causing #829 and #839.