openDialogLink has public `open()` method
@jophis here’s my solution:

1. We were containing `OpenDialogLink` within a dropdown, which would
close after clicking, so all components inside of it would get closed
(and probably dismounted) every time. So, we need to move this to
somewhere more top-level, safest place would be at the very top of
`.planView`. It won’t have any children.

2. Give it a ref property “createPlanTemplateDialog”, so we can call
internal methods on it from planTab.

3. Have your MenuItem’s onClick call the `open()` method, using the
ref. This will cause your `OpenDialogLink` to open the dialog - as if
it was clicked normally. For more examples of this, just do a big
search for “@refs”.